+++ CHANGELOG.md
@@ -16,9 +16,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
-- http-out-response header obfuscation attempt in case of nil response
 
 ### Security
+
+## [v1.56.0] - 2022-05-24
+
+### Fixed
+
+- http-out-response header obfuscation attempt in case of nil response
 
 ## [v1.55.0] - 2022-05-11
 
@@ -739,7 +744,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.55.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.56.0...HEAD
+[v1.56.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.55.0...v1.56.0
 [v1.55.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.54.0...v1.55.0
 [v1.54.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.53.1...v1.54.0
 [v1.53.1]: https://github.com/gesundheitscloud/go-svc/compare/v1.53.0...v1.53.1